feat(owner-creation): implement nested writes for simplified Owner creation
This commit introduces a change in how Owners are created. The process now leverages Prisma's nested writes, streamlining the logic. When a user creates an Owner, a User is concurrently created in the background, along with a UserRole and an Owner associated with the newly created User. It simplifies the Owner creation process.

Note: Owners created through this method will always have the role of "CUSTOMER." Only Users with the role of "ADMIN" or "EMPLOYEE" may create an Owner through this process.

Additionally, when a user registers themselves, a User is created, and in a nested write, an Owner and UserRole are also created. This enhancement simplifies the user registration logic.

Note: Users created through this method will always have a role of "CUSTOMER." Only Customers can create users for themselves.