feat(owners): implemented delete functionality for owners
Implement owner/user delete server action. Modify ui, so that the user gets
prompted to confirm if they really want to delete the user. Change prisma
schema to allow for cascading deletes. That way we only need the userId and
we delete the user and any associated records, that can be of type Owner,
Vet, UserRole, etc.

Refactor the form modal so it can be reused more easily.