feat(vets-page): add vet's page and list component
Add vet's page and list. Modify form slice to differentiate between owner
form and vet form.

Modify form container to allow for different headers depending on which
type of user we're adding or editing.

Modify delete form to remove the correct type of entity, owner or vet.

Rename a few files for more clarity and ease of navigation.

Add zod schema for vet.