feat(vets): add vet's form
Add vet's form and fix editing bugs. Before when we added an owner or vet,
and imediately tried to edit it, it wasn't working because we weren't returning
the id of the model, so we were adding it to the redux store without the id.
So when we tried to get it, there was no id for that newly added model
and it threw an error. Now, we just include the vet or owner in the prisma
create function and then return it in the action as well and get the id that way.