fix(layout): fix mobile responsiveness on tables, sidebar and container
Fix tables so that each column progressively appears depending on screen size.

Fix sidebar so that instead of shrinking when closed on small devices,
it completly disappears. Then when the user opens the sidebar, it shows
fully over the content with an overlay behind.

Fix container paddings, so there is no padding on small devices, whether
the sidebar is open or not.