refactor(form): modify forms to use shadcn controlled components
Implement new component ControlledComboBox that is used to select an Owner
using shadcn/ui's Popover and Command in the background.

Modify pet and vet forms to use shadcn FormField in order to simplify the validations flow.

Add new Command component from shadcn/ui, that is used in combination with a
Popover to create a ComboBox component.

Renamed file with server actions from "index.ts" to "actions.ts" for clarity.

Small code cleanup.