fix(blob-upload): move file upload to the client
Move fetch call to the client in the since vercel doesn't allow for
files to be uploaded from server actions.

Fix owner select combobox, there was some issue related to the hover effect.
Changed the value prop of the CommandItem component to something unique, like the id of the owner.