feat(auth): add email confirmation flow
Add email confirmation flow for users created through the register page
and users created through the owner or vet creation forms.
Users created through the register page are sent a link and all they have
to do is click the link and the account should be activated.
Users created through the vet or owners page are sent a link and when they
navigate to that link they are prompted to create a password.

Add a VericationToken model and table to the db, this model is used to
temporarly save the verification email tokens.

Add resend package and config.

Add uuid package to generate the tokens.