refactor(blob): modify image upload logic
Create a hook "useImageUpload" that gives us a function "upload", that takes
a file and handles all the logic necessary to upload a file on the browser.
It first fetches "/api/upload", in this endpoint we create a pre-signed post
and return the url and fields, then back in upload we fetch the url with
the fields and the file appended as FormData. Then we return the complete
url with the file name (key) to be save with our entity.

Move redux hooks into redux folder and update imports.