refactor(blob): implement useImageUpload for pets and vets
Modify pets and vets image upload flow to use the hook I created
"useImageUpload". Add deleteImage method to the useImageUpload hook, so
we can more easily delete images from the browser. Very usefull for when
the entity being edited already had an image and the user changes it, we want to delete the old one, or if we already uploaded an image, but there was a problem updating the user.

Implement image delete when a user or a pet is deleted.