fix(blob): upload and delete image logic
Before, when the user selected an image in the form, there was a possibility
that we would delete the old/current image before updating the entity
with the new imageUrl, meaning if the update failed the entity would remain the same,
but their imageUrl would no longer be valid, as the image was deleted from the bucket.

Now, we wait till after the entity was updated successfully, then we check
if the user had an image and if their new image was uploaded and only then
we delete the old image.