feat(time-picker): implement time range picker
Implement a time range picker that works with shadcn/ui selectors, it is
composed of two separate time pickers. Each time picker is in increments
of 30 minites. When we select a start time, the values on the end time selector automatically change, so we are only allowed to pick times that are after
the time selected on the start time selector.