feat(scheduler): add scheduler form
Add scheduler form base with validations for the startTime and endTime.

Make TimePicker and TimeRangePicker a controlled component, using FormField.
Add additional logic to the timepicker so that if the user selects a startTime
that is bigger than the current selected endTime, we change the value of
endTime automatically to 30 minutes after startTime.