feat(scheduler): appointment form auto-population
Implement form auto-population when the user clicks a cell on the scheduler.
If it's an empty cell the on the month view, the time-pickers will receive a default
starTime and endtime, all the other fields will be empty. If it's an empty
cell on the day or week view, the time-pickers will receive the time of the
cell the user selected, all the other fields will be empty.

If the user clicked an appointment, the form will populated with all the correct
data.

Modified TimePicker and TimeRangePicker, so that the TimeRangePicker, controls
the logic including the creation of the timeSlots for starTime and endtime,
which are then passdown to the respective TimePicker.

Add AppoinmentData to the scheduler context, so we can access all the data
about an appointment in the appointment form.