feat(scheduler): implement resize and drag updates
Implement the rezise and drag updates, so that it updates the appointment
on the database.

Add toast messages to all appointment managment actions.

Fix return from updateAppoinment, it was returning false when it should
be returning true.

Fix bug when creating and then updating an appointment without refreshing
the page. the newly created appointments were being added without an id
which resulted in the update failing.