feat(chat): add db actions to create messages and conversations
Add prisma schemas models message and conversation and relationships with
user. Add actions to create conversations and messages, and retrieve the
conversations of the logged in users, and messages of a particular conversation.
Implement the creation of messages nad retrieval of existing conversations.

Devide chat ui into smaller components, so the logic and fetching of data
can be split in a more organized way. With the conversations, I'm fetching
all the conversations of a particular user, then each conversation component
is fetching the other user, meaning the user that is not logged and is a
part of the conversation, and displaying that data.

Modify ChatList to take in a list of conversations, instead of owners.