feat(chat): add create conversation menu
Add a component that consists of a popup menu with a list of users that
doesn't contain the logged in user. When the user clicks a user from the list a
conversation is created and the user can now send messages in that conversation.

Conversations are now fetched in the ChatLayout component. I changed the
getConversations action, so that we return the list of Conversations, but now we
filter the userConversations array in each one, so that we only get the
userConversation where the userId that doesn't match the logged in user's id.
I created an interface to handle this new return type.
This way we can simply pass a conversation to the ChatConversation component
and use the user in that conversation without having to fetch the user per
ChatConversation rendered.