feat(chat): filter conversation per user role
Modify get conversations action, so that we get also the roles of each user.
Implement a hook useFilterConversations that takes the conversations and returns
two functions: getCustomerConversations and getEmployeeConversations. Each
of these functions returns an array of conversations based on the role of the user
the conversation is with.
Modify ChatList so that now it uses a Command with a Command group for each
type of conversation: Owner and Vet. With this the user can now search for
conversations using the name of a user and also it is easier to differentiate
the types of conversation.

Add a server action to get only the users of role "EMPLOYEE" and Modify
getUsers so that now it gets only users of role "CUSTOMER". With this we can
differentiate between different roles in the the add conversation popup,
which makes it easier for the user to navigate the menu and know which user
they are adding. If a user is a customer then, they can only add conversations
with users of role "EMPLOYEE".