refactor(chat): move user filterting into a hooks
Move the logic of filtering the users in the new conversation popup into
a hook useFilterUsers that returns a function filterUsers, an array of
customer users, an array of employee users, and another function removeUser.
filterUsers takes an array of users and separates the users per role into
two different arrays and returns them.
removeUser receives the selected user's id and removes that user from the
array of all users and calls filterUsers after, this function is called
after a new conversation is created.

Move logic to get all users into a route handler. I'm planning on moving most
server actions into route handlers, as I feel this approach is cleaner.