feat(dashboard): create dashboard page and grid
Create dashboard page and grid. Create dashboard messages component, that
shows all the latest messages.

Create extended types based on the prisma generated types that can be used
to receive results of queries to tables with relations.

Update components and hooks to now use the extended types.

Modify getConversations action so that it filters the users automatically
in the query.

Add not found page.