fix(dashboard): fix getUpcomingAppointments server action
Modify getUpcomingAppointments to differentiate between the type of user
getting the upcomging appointments.
If the user is of role "ADMIN", we get all the appointments.
If the user is of role "VET", we get only the appointments relevant to that vet.
If the user is of role owner, we get all the ids of that owner's pets,
and then we get all the appointments of those pets.