feat(auth): add reset password flow
Create a send reset-password email page and form, when the form
is submitted we send an email with a link to a reset password page with
a token.

Create reset-password page, the link to this page is sent in the email.
The page contains a change-password form, which is the same form we use
in the confirm-email page. We added a flag to this component, so that we
can differentiate between a confirm-email and a reset-password and display
the correct messages.

Restructure folders and move code from confirm-email into the confirm email page.
Delete confirm-email folder and component and move change-password-form to a new
folder, so the structure is a bit simpler.