feat(auth): implement role checking in sidebar
Extend SidebarItemProps into new interface, that also contains a
rolesAllowed string array with all the roles that are allowed to see
a specific page.
Implement a function that takes a rolesToCheck array and checks to see
if any of the roles of the current logged in user match and renders the
sidebarItem(or not) based on that.