feat(auth): add a role gate to manage access control
Add a component RoleGate that takes in as props an array of rolesAllowed
and a children react node. Then it checks to see if any of those rolesAllowed
match with the roles of the logged in user, if no roles match it returns
an unauthorized page, otherwise, it renders the children.