refactor(sidebar): modify rendering logic for sidebar items
Modify SideBarItemProps to add a string property pathName, which we now
use to determine the active page. Remove unnecessary flag displayTitle.

Modify constants so that now instead of having one array of SideBarItemProps
and display(or not) the sidebar item, we have two: one for owners and one for
employees, and we display this array based on the roles of the logged in user.
Then, all "common" sidebar items we display directly.

Add a MyPetsPage, where we will display all the pets related to an owner,
in this page owners will be able to manage their own pets.