chore(seed): add seed data and modify role names
Modify prisma schema so that the roles "CUSTOMER" and "EMPLOYEE" are now called
"OWNER" and "VET" respectively, this made it more clear and simple. Replace all instances of
"CUSTOMER" and "EMPLOYEE".

Create a file seed-data.ts that has some hardcoded owner, vet and pets objects that
we use to seed the database. In this file we also created a function that generates
programatically a specific number of appointments with random values that follow
some rules.

Remove all console.logs from the project because I had accumulated many over the
development of the project and it was getting harder to debug stuff.